project add

+++ src/com/examination_system/dao/impl/UserDaoImpl.java
@@ -20,6 +20,12 @@ public class UserDaoImpl implements UserDao {
 
     private static final String SELECT_COLUMNS = "select id, name, user_no as userNo, user_type as userType, password, is_delete as isDelete ";
 
+    /**
+     * 根据主键id查询用户实体类
+     *
+     * @param id 主键id
+     * @return User
+     */
     @Override
     public User getById(Long id) {
         String sql = SELECT_COLUMNS + "  where id = ?";
@@ -28,6 +34,12 @@ public class UserDaoImpl implements UserDao {
         return CRUDTemplate.executeQuery(sql, new BeanHandler<>(User.class), params);
     }
 
+    /**
+     * 添加用户
+     *
+     * @param user 用户实体类
+     * @return int
+     */
     @Override
     public int add(User user) {
         String sql = "INSERT INTO user (name,user_no,user_type,password,is_delete) VALUES (?,?,?,?,0)";
@@ -39,6 +51,12 @@ public class UserDaoImpl implements UserDao {
         return CRUDTemplate.executeUpdate(sql, param);
     }
 
+    /**
+     * 修改用户
+     *
+     * @param user 用户实体类
+     * @return int
+     */
     @Override
     public int update(User user) {
         String sql = "UPDATE user set name = ?,user_no = ?,user_type = ?, password = ? WHERE id = ?";
@@ -52,13 +70,19 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public int delete(int id) {
+    public int delete(Long id) {
         String sql = "UPDATE user set is_delete = 1, password WHERE id = ?";
         List<Object> param = new ArrayList<>();
         param.add(id);
         return CRUDTemplate.executeUpdate(sql, param);
     }
 
+    /**
+     * 根据主键id删除用户
+     *
+     * @param id 主键id
+     * @return int
+     */
     @Override
     public User getByName(String name) {
         String sql = SELECT_COLUMNS + "  where name = ?";
@@ -67,6 +91,12 @@ public class UserDaoImpl implements UserDao {
         return CRUDTemplate.executeQuery(sql, new BeanHandler<>(User.class), param);
     }
 
+    /**
+     * 根据用户名查询用户实体类
+     *
+     * @param name 用户名
+     * @return User
+     */
     @Override
     public List<User> queryList(UserQuery query) {
         String sql = SELECT_COLUMNS + "  where  1=1 ";
@@ -80,6 +110,12 @@ public class UserDaoImpl implements UserDao {
         return CRUDTemplate.executeQuery(sql, new BeanListHandler<>(User.class), param);
     }
 
+    /**
+     * 根据查询条件分页用户实体类集合
+     *
+     * @param query 主键id
+     * @return List<User>
+     */
     @Override
     public List<User> queryListByPage(UserQuery query) {
         String sql = SELECT_COLUMNS + "  where  1=1 ";
@@ -93,6 +129,12 @@ public class UserDaoImpl implements UserDao {
         return CRUDTemplate.executeQuery(sql, new BeanListHandler<>(User.class), param);
     }
 
+    /**
+     * 根据查询条件分页用户实总条数
+     *
+     * @param query 主键id
+     * @return Integer
+     */
     @Override
     public Integer queryPageCount(UserQuery query) {
         String sql = SELECT_COLUMNS + "  where  1=1 ";
@@ -102,6 +144,13 @@ public class UserDaoImpl implements UserDao {
         return CRUDTemplate.executeQuery(sql, new BeanHandler<>(Integer.class), param);
     }
 
+    /**
+     * 配置参数
+     *
+     * @param sql   sql
+     * @param param 参数列表
+     * @param query 查询条件
+     */
     private void setParam(String sql, List<Object> param, UserQuery query) {
         if (query.getName() != null && query.getName().trim().length() > 0) {
             sql = sql + " and name = ?";
